+++ UniWatch/DataAccess/IStudentManager.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using UniWatch.Models;
 
-namespace UniWatch.Managers
+namespace UniWatch.DataAccess
 {
     public interface IStudentManager : IDisposable
     {
